requirements.txt
fastapi
uvicorn
kafka-python
pymongo
requests
python-dotenv
